personal_data/requirements.txt
django >= 2.2.13, < 3.0
gunicorn
django-heroku
djangorestframework
apscheduler
whitenoise